--- src/main/java/com/kamesuta/mc/autoinput/gui/InputKey.java
@@ -1,5 +1,0 @@
-package com.kamesuta.mc.autoinput.gui;
-
-public class InputKey {
-
-}
